refact add destination and offers model
