+++ src/model/destination-model.js
@@ -1,0 +1,13 @@
+import { mockDestinations } from '../mock/destinations.js';
+
+export default class DestinationModel {
+  #destinations = mockDestinations;
+
+  get destinations() {
+    return this.#destinations;
+  }
+
+  getDestinationById(id) {
+    return this.#destinations.find((destination) => destination.id === id);
+  }
+}
